--- Kit/FluentValidation.Base/RuleBuilderExtensions.cs
@@ -1,15 +1,0 @@
-﻿//using BlazorDevKit.FluentValidation.Base;
-
-//namespace FluentValidation;
-//public static class RuleBuilderExtensions
-//{
-//    public static IRuleBuilderOptions<T, TProperty> NotEmptyTrackRequired<T, TProperty>(this IRuleBuilder<T, TProperty> ruleBuilder)
-//    {
-//        return ruleBuilder.SetValidator(new NotEmptyTrackRequiredValidator<T, TProperty>());
-//    }
-
-//    public static IRuleBuilderOptions<T, TProperty> NotNullTrackRequired<T, TProperty>(this IRuleBuilder<T, TProperty> ruleBuilder)
-//    {
-//        return ruleBuilder.SetValidator(new NotNullTrackRequiredValidator<T, TProperty>());
-//    }
-//}
