+++ Kit/FluentValidation/BdkFluentValidator.razor
@@ -1,2 +1,1 @@
 ﻿@typeparam TRequest where TRequest: class
-@ChildContent
