+++ Kit/FluentValidation/BlazorDevKit.FluentValidation.csproj
@@ -30,7 +30,4 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="8.0.8" />
 	</ItemGroup>
-	<ItemGroup>
-	  <ProjectReference Include="..\FluentValidation.Base\BlazorDevKit.FluentValidation.Base.csproj" />
-	</ItemGroup>
 </Project>
